+++ SoundFontsFramework/SoundFontsFramework/Resources/Changes.md
@@ -6,6 +6,11 @@ The parsing is really simplistic:
 - always put most-recent changes at the top of the file (versions in descending order)
 - if a line begins with '* ' then it is a change entry to show to the user. The change *must* be all on one line (keep it short and sweet)
 
+# 2.28.0
+
+* Adopt MIDI 2.0 CoreMIDI API when available (iOS 14+)
+* Fix bug when processing MIDI 1.0 running status
+
 # 2.27.4
 
 * Simplify shift A4 behavior.